+++ Server/Utils/wrapper.h
@@ -1,6 +1,7 @@
 #ifndef WRAPPER_H
 #define WRAPPER_H
 
+#include <sys/types.h>
 #include <stdint.h>
 #include <stddef.h>
 
@@ -9,8 +10,8 @@ typedef struct timeval timeval;
 int Connection_SetTimeout(int socket, int seconds);
 
 // return n bytes read, 0 on close, -1 on error
-int Connection_Recv(int socket, uint8_t *buffer, size_t length);
+ssize_t Connection_Recv(int socket, uint8_t *buffer, size_t length);
 // return n bytes written, -1 on error
-int Connection_Send(int socket, const char *buffer, size_t length);
+ssize_t Connection_Send(int socket, const char *buffer, size_t length);
 
 #endif
